backend/requirements.txt
Flask
Flask-CORS
gunicorn
tensorflow
scikit-learn
joblib
Pillow
numpy
